refactor: remove polling from races and increase cache interval
